+++ lib/drip.dart
@@ -1,3 +1,4 @@
+export 'src/async_drip.dart';
 export 'src/computed_drip.dart';
 export 'src/drip_core.dart';
 export 'src/drip_observer.dart';